Issue #202: Reduce the fileLength of InputMutation14
