+++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests21/DiffMutationTests21.diff
@@ -1,0 +1,137 @@
+--- src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests21/InputMutationTests21.java
++++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests21/OutputMutationTests21.java
+@@ -7,63 +7,47 @@
+ */
+ package org.checkstyle.autofix.recipe.useenhancedswitch.mutationtests21;
+ 
+-public class InputMutationTests21 {
++public class OutputMutationTests21 {
+     enum Color { RED, GREEN, BLUE }
+ 
+     void testDifferentVariableAssignment(int x) {
+         int y;
+         int z;
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1:
+-                y = 1;
+-                break;
+-            case 2:
+-                z = 1;
+-                break;
+-            default:
+-                y = 0;
+-                break;
++        switch (x) {
++            case 1 -> y = 1;
++            case 2 -> z = 1;
++            default -> y = 0;
+         }
+     }
+ 
+     void testEmptyCase(int x) {
+         int y;
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1:
+-                y = 1;
+-                break;
+-            case 2:
+-                break;
+-            default:
+-                y = 0;
+-                break;
++        switch (x) {
++            case 1 -> y = 1;
++            case 2 -> {}
++            default -> y = 0;
+         }
+     }
+ 
+     enum OtherColor { RED, GREEN }
+     void testMultipleEnums(Color c, OtherColor oc) {
+-        int x;
+-        switch (c) { // violation 'Switch can be replaced with enhanced switch.'
+-            case RED: x = 1; break;
+-            case GREEN: x = 2; break;
+-            case BLUE: x = 3; break;
+-        }
++        int x = switch (c) {
++            case RED -> 1;
++            case GREEN -> 2;
++            case BLUE -> 3;
++        };
+     }
+ 
+     void testContinueInCase(int x) {
+         for (int i = 0; i < 10; i++) {
+-            switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-                case 1:
++            switch (x) {
++                case 1 -> {
+                     if (x > 0) {
+                         continue;
+                     }
+-                    System.out.println("one");
+-                    break;
+-                case 2:
+-                    System.out.println("two");
+-                    break;
+-                default:
+-                    break;
++                    System.out.println("one");}
++                case 2 -> System.out.println("two");
++                default -> {}
+             }
+         }
+     }
+@@ -71,38 +55,29 @@
+     void testLabeledBreakInCase(int x) {
+         outer:
+         for (int i = 0; i < 10; i++) {
+-            switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-                case 1:
++            switch (x) {
++                case 1 -> {
+                     if (x > 0) {
+                         break outer;
+                     }
+-                    System.out.println("one");
+-                    break;
+-                case 2:
+-                    System.out.println("two");
+-                    break;
+-                default:
+-                    break;
++                    System.out.println("one");}
++                case 2 -> System.out.println("two");
++                default -> {}
+             }
+         }
+     }
+ 
+     void testThrowInAssignment(int x) {
+-        int y;
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1:
+-                y = 1;
+-                break;
+-            default:
+-                throw new RuntimeException("bad");
+-        }
++        int y = switch (x) {
++            case 1 -> 1;
++            default -> throw new RuntimeException("bad");
++        };
+     }
+ 
+     void testNoNewlineIndent(int x) {
+-        int y;
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1: y = 1; break;
+-            default: y = 0; break;
+-        }
++        int y = switch (x) {
++            case 1 -> 1;
++            default -> 0;
++        };
+     }
+ }
