+++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests22/DiffMutationTests22.diff
@@ -1,0 +1,133 @@
+--- src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests22/InputMutationTests22.java
++++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests22/OutputMutationTests22.java
+@@ -7,98 +7,84 @@
+ */
+ package org.checkstyle.autofix.recipe.useenhancedswitch.mutationtests22;
+ 
+-public class InputMutationTests22 {
++public class OutputMutationTests22 {
+     void testNegativeControlFlow(int x) {
+         int y;
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1:
++        switch (x) {
++            case 1 -> {
+                 if (x > 0) return;
+-                y = 1;
+-                break;
+-            default:
+-                y = 0;
+-                break;
++                y = 1;}
++            default -> y = 0;
+         }
+     }
+ 
+     void testNegativeAssignment(int x) {
+         int[] arr = new int[1];
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1:
+-                arr[0] = 1;
+-                break;
+-            default:
+-                arr[0] = 0;
+-                break;
++        switch (x) {
++            case 1 -> arr[0] = 1;
++            default -> arr[0] = 0;
+         }
+     }
+ 
+     enum SmallEnum { A, B }
+     void testNegativeEnum(SmallEnum e) {
+         int x;
+-        switch (e) { // violation 'Switch can be replaced with enhanced switch.'
+-            case A: x = 1; break;
++        switch (e) {
++            case A -> x = 1;
+         }
+     }
+ 
+     void testNegativeAssignmentNoAssign(int x) {
+         int y;
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1: System.out.println(1); break;
+-            default: y = 0; break;
++        switch (x) {
++            case 1 -> System.out.println(1);
++            default -> y = 0;
+         }
+     }
+ 
+     int testNegativeReturnNoDefault(SmallEnum e) {
+-        switch (e) { // violation 'Switch can be replaced with enhanced switch.'
+-            case A: return 1;
++        switch (e) {
++            case A -> { return 1;}
+         }
+         return 0;
+     }
+ 
+     int testNegativeReturnControlFlow(int x) {
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1:
++        switch (x) {
++            case 1 -> {
+                 if (x > 0) {
+                     for (int i=0; i<10; i++) {
+                         if (i == 5) break;
+                     }
+                 }
+-                return 1;
+-            default: return 0;
++                return 1;}
++            default -> { return 0;
++            }
+         }
+     }
+ 
+     void testAssignmentFallthrough(int x) {
+-        int y;
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1:
+-            case 2: y = 1; break;
+-            default: y = 0; break;
+-        }
++        int y = switch (x) {
++            case 1, 2 -> 1;
++            default -> 0;
++        };
+     }
+ 
+     void testAssignmentBlock(int x) {
+-        int y;
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1: {
++        int y = switch (x) {
++            case 1 -> {
+                 System.out.println(1);
+-                y = 1;
+-                break;
++                yield 1;
+             }
+-            default: y = 0; break;
+-        }
++            default -> 0;
++        };
+     }
+ 
+     void testNegativeAssignmentNonIdentifier(int x) {
+         int y;
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1:
+-                this.y = 1; // Not y = 1;
+-                break;
+-            default:
+-                y = 0;
+-                break;
++        switch (x) {
++            case 1 -> this.y = 1;
++            default -> y = 0;
+         }
+     }
+
